refactor(pdf): isolate display equation evidence
